# openapi.yaml
components:
    schemas:
...
        Category:
            properties:
                CreatedAt:
                    format: date-time
                    readOnly: true
                    type: string
                CustomAvatar:
                    type: string
                CustomCSS:
                    type: string
                CustomHeader:
                    type: string
                CustomMRQLResult:
                    type: string
                CustomSidebar:
                    type: string
                CustomSummary:
                    type: string
                Description:
                    type: string
                Groups:
                    items:
                        $ref: '#/components/schemas/GroupPartial'
                    type: array
                ID:
                    readOnly: true
                    type: integer
                MetaSchema:
                    type: string
                Name:
                    type: string
                UpdatedAt:
                    format: date-time
                    readOnly: true
                    type: string
                guid:
                    nullable: true
                    type: string
                sectionConfig:
                    additionalProperties: true
                    description: Arbitrary JSON data
                    type: object
            type: object
        CategoryEditor:
            properties:
                CustomAvatar:
                    type: string
                CustomCSS:
                    type: string
                CustomHeader:
                    type: string
                CustomMRQLResult:
                    type: string
                CustomSidebar:
                    type: string
                CustomSummary:
                    type: string
                Description:
                    type: string
                ID:
                    type: integer
                MetaSchema:
                    type: string
                Name:
                    type: string
                SectionConfig:
                    type: string
            type: object
...
        ExportEstimate:
            properties:
                counts:
                    $ref: '#/components/schemas/Counts'
                danglingByKind:
                    additionalProperties: true
                    type: object
                estimatedBytes:
                    type: integer
                uniqueBlobs:
                    type: integer
            type: object
...
        ExportRequest:
            properties:
                fidelity:
                    $ref: '#/components/schemas/ExportFidelity'
                gzip:
                    type: boolean
                relatedDepth:
                    type: integer
                rootGroupIds:
                    items:
                        type: integer
                    type: array
                schemaDefs:
                    $ref: '#/components/schemas/ExportSchemaDefs'
                scope:
                    $ref: '#/components/schemas/ExportScope'
            type: object
...
        NoteType:
            properties:
                CreatedAt:
                    format: date-time
                    readOnly: true
                    type: string
                CustomAvatar:
                    type: string
                CustomCSS:
                    type: string
                CustomHeader:
                    type: string
                CustomMRQLResult:
                    type: string
                CustomSidebar:
                    type: string
                CustomSummary:
                    type: string
                Description:
                    type: string
                ID:
                    readOnly: true
                    type: integer
                MetaSchema:
                    type: string
                Name:
                    type: string
                Notes:
                    items:
                        $ref: '#/components/schemas/NotePartial'
                    type: array
                SectionConfig:
                    additionalProperties: true
                    description: Arbitrary JSON data
                    type: object
                UpdatedAt:
                    format: date-time
                    readOnly: true
                    type: string
                guid:
                    nullable: true
                    type: string
            type: object
        NoteTypeEditor:
            properties:
                CustomAvatar:
                    type: string
                CustomCSS:
                    type: string
                CustomHeader:
                    type: string
                CustomMRQLResult:
                    type: string
                CustomSidebar:
                    type: string
                CustomSummary:
                    type: string
                Description:
                    type: string
                ID:
                    type: integer
                MetaSchema:
                    type: string
                Name:
                    type: string
                SectionConfig:
                    type: string
            type: object
...
        QueryEditor:
            properties:
                Description:
                    type: string
                ID:
                    type: integer
                Name:
                    type: string
                Template:
                    type: string
                Text:
                    type: string
            type: object
...
        SavedMRQLQuery:
            properties:
                createdAt:
                    format: date-time
                    readOnly: true
                    type: string
                description:
                    type: string
                id:
                    readOnly: true
                    type: integer
                name:
                    type: string
                query:
                    type: string
                updatedAt:
                    format: date-time
                    readOnly: true
                    type: string
            type: object
        SavedMRQLQueryPartial:
            properties:
                id:
                    type: integer
                name:
                    type: string
            type: object
...
info:
    title: mahresources
    version: "1.0"
openapi: 3.0.0
paths:
...
    /v1/auth/login:
        post:
            description: Exchanges a username/password for a session cookie. Only meaningful when auth is enabled.
            operationId: login
            requestBody:
                content:
                    application/json:
                        schema:
                            properties:
                                password:
                                    type: string
                                username:
                                    type: string
                            type: object
                    application/x-www-form-urlencoded:
                        schema:
                            properties:
                                password:
                                    type: string
                                username:
                                    type: string
                            type: object
                required: true
            responses:
                "200":
                    content:
                        application/json: {}
                    description: Successful response
            summary: Authenticate and start a session
            tags:
                - auth
...
    /v1/category:
        post:
            operationId: createOrUpdateCategory
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/CategoryEditor'
                    application/x-www-form-urlencoded:
                        schema:
                            $ref: '#/components/schemas/CategoryEditor'
                required: true
            responses:
                "200":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Category'
                    description: Successful response
            summary: Create or update a category
            tags:
                - categories
...
    /v1/groups/export/estimate:
        post:
            description: Walks the requested scope without writing a tar; returns counts, unique blob count, dangling reference summary.
            operationId: estimateGroupExport
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/ExportRequest'
                required: true
            responses:
                "200":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ExportEstimate'
                    description: Successful response
            summary: Estimate the size and shape of a proposed group export
            tags:
                - exports
...
    /v1/mrql:
        post:
            description: |-
                Parses, validates, translates, and executes a Mahresources Query Language (MRQL) query.

                Request body fields:
                  - query   (string, required) — the MRQL source
                  - limit   (integer)          — items per bucket (grouped) or total items
                  - buckets (integer)          — buckets per page (grouped mode only)
                  - page    (integer)          — 1-based page number
                  - offset  (integer)          — explicit cursor offset (takes precedence over page)

                Query parameter:
                  - render (0 or 1) — when 1, populates each result row's RenderedHTML using
                    the entity's CustomMRQLResult template.
            operationId: executeMRQL
            parameters:
                - description: Set to 1 to render CustomMRQLResult templates
                  in: query
                  name: render
                  schema:
                    type: integer
            responses:
                "200":
                    content:
                        application/json: {}
                    description: Successful response
            summary: Execute an MRQL query
            tags:
                - mrql
...
    /v1/mrql/saved:
        get:
            description: 'Without `id`: paginated list of saved queries (pass `all=1` for the full set). With `id`: returns a single saved query.'
            operationId: listSavedMRQLQueries
            parameters:
                - description: Fetch a single saved query by ID
                  in: query
                  name: id
                  schema:
                    type: integer
                - description: Set to 1 to disable pagination
                  in: query
                  name: all
                  schema:
                    type: integer
                - description: Page number for pagination
                  in: query
                  name: page
                  schema:
                    default: 1
                    type: integer
            responses:
                "200":
                    content:
                        application/json:
                            schema:
                                items:
                                    $ref: '#/components/schemas/SavedMRQLQueryPartial'
                                type: array
                    description: Successful response
            summary: List or fetch saved MRQL queries
            tags:
                - mrql
        post:
            description: |-
                Request body fields:
                  - name        (string, required)
                  - query       (string, required) — MRQL source
                  - description (string)
            operationId: createSavedMRQLQuery
            responses:
                "200":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/SavedMRQLQuery'
                    description: Successful response
            summary: Create a saved MRQL query
            tags:
                - mrql
        put:
            operationId: updateSavedMRQLQuery
            parameters:
                - in: query
                  name: id
                  required: true
                  schema:
                    type: integer
            responses:
                "200":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/SavedMRQLQuery'
                    description: Successful response
            summary: Update a saved MRQL query
            tags:
                - mrql
...
    /v1/mrql/saved/run:
        post:
            description: Runs a previously saved MRQL query. Accepts either `id` or `name` to identify the saved query, plus the same pagination params as /v1/mrql.
            operationId: runSavedMRQLQuery
            parameters:
                - description: Saved query ID
                  in: query
                  name: id
                  schema:
                    type: integer
                - description: Saved query name (fallback if id not found)
                  in: query
                  name: name
                  schema:
                    type: string
                - in: query
                  name: limit
                  schema:
                    type: integer
                - in: query
                  name: page
                  schema:
                    type: integer
                - in: query
                  name: buckets
                  schema:
                    type: integer
                - in: query
                  name: offset
                  schema:
                    type: integer
                - description: Set to 1 to render CustomMRQLResult templates
                  in: query
                  name: render
                  schema:
                    type: integer
            responses:
                "200":
                    content:
                        application/json: {}
                    description: Successful response
            summary: Execute a saved MRQL query by id or name
            tags:
                - mrql
...
    /v1/note/noteType:
        post:
            operationId: createNoteType
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/NoteTypeEditor'
                    application/x-www-form-urlencoded:
                        schema:
                            $ref: '#/components/schemas/NoteTypeEditor'
                required: true
            responses:
                "200":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/NoteType'
                    description: Successful response
            summary: Create a new note type
            tags:
                - notes
...
    /v1/plugin/disable:
        post:
            operationId: disablePlugin
            parameters:
                - description: Plugin name to disable
                  in: query
                  name: name
                  required: true
                  schema:
                    type: string
            responses:
                "200":
                    content:
                        application/json: {}
                    description: Successful response
            summary: Disable a plugin
            tags:
                - plugins
    /v1/plugin/enable:
        post:
            operationId: enablePlugin
            parameters:
                - description: Plugin name to enable
                  in: query
                  name: name
                  required: true
                  schema:
                    type: string
            responses:
                "200":
                    content:
                        application/json: {}
                    description: Successful response
            summary: Enable a plugin
            tags:
                - plugins
    /v1/plugin/purge-data:
        post:
            operationId: purgePluginData
            parameters:
                - description: Plugin name to purge data for
                  in: query
                  name: name
                  required: true
                  schema:
                    type: string
            responses:
                "200":
                    content:
                        application/json: {}
                    description: Successful response
            summary: Purge all data for a plugin
            tags:
                - plugins
    /v1/plugin/settings:
        post:
            operationId: updatePluginSettings
            parameters:
                - description: Plugin name
                  in: query
                  name: name
                  required: true
                  schema:
                    type: string
            responses:
                "200":
                    content:
                        application/json: {}
                    description: Successful response
            summary: Update plugin settings
            tags:
                - plugins
...
    /v1/query:
        get:
            operationId: getQuery
            parameters:
                - in: query
                  name: id
                  required: true
                  schema:
                    type: integer
            responses:
                "200":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Query'
                    description: Successful response
            summary: Get a specific query
            tags:
                - queries
        post:
            operationId: createOrUpdateQuery
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/QueryEditor'
                    application/x-www-form-urlencoded:
                        schema:
                            $ref: '#/components/schemas/QueryEditor'
                required: true
            responses:
                "200":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Query'
                    description: Successful response
            summary: Create or update a query
            tags:
                - queries
...
    /v1/query/run:
        post:
            operationId: runQuery
            parameters:
                - in: query
                  name: id
                  schema:
                    type: integer
                - description: Name of the query to run (alternative to id)
                  in: query
                  name: name
                  schema:
                    type: string
            responses:
                "200":
                    description: Successful response
            summary: Run a saved query
            tags:
                - queries
...
    /v1/relation:
        post:
            operationId: createOrUpdateRelation
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/GroupRelationshipQuery'
                    application/x-www-form-urlencoded:
                        schema:
                            $ref: '#/components/schemas/GroupRelationshipQuery'
                required: true
            responses:
                "200":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/GroupRelation'
                    description: Successful response
            summary: Create or edit a group relation instance
            tags:
                - relations
...
    /v1/resource/removeSeries:
        post:
            operationId: removeResourceFromSeries
            parameters:
                - in: query
                  name: id
                  required: true
                  schema:
                    type: integer
            responses:
                "200":
                    description: Successful response
            summary: Remove a resource from its series
            tags:
                - series
...
    /v1/resourceCategory:
        post:
            operationId: createOrUpdateResourceCategory
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/ResourceCategoryEditor'
                    application/x-www-form-urlencoded:
                        schema:
                            $ref: '#/components/schemas/ResourceCategoryEditor'
                required: true
            responses:
                "200":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ResourceCategory'
                    description: Successful response
            summary: Create or update a resource category
            tags:
                - resourceCategories
...
    /v1/search:
        get:
            operationId: globalSearch
            parameters:
                - in: query
                  name: q
                  schema:
                    type: string
                - in: query
                  name: limit
                  schema:
                    type: integer
                - explode: true
                  in: query
                  name: types
                  schema:
                    items:
                        type: string
                    type: array
                  style: form
                - description: Page number for pagination
                  in: query
                  name: page
                  schema:
                    default: 1
                    type: integer
            responses:
                "200":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/GlobalSearchResponse'
                    description: Successful response
            summary: Global search across all entities
            tags:
                - search
    /v1/series:
        get:
            operationId: getSeries
            parameters:
                - in: query
                  name: id
                  required: true
                  schema:
                    type: integer
            responses:
                "200":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Series'
                    description: Successful response
            summary: Get a specific series
            tags:
                - series
        post:
            operationId: updateSeries
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/SeriesEditor'
                    application/x-www-form-urlencoded:
                        schema:
                            $ref: '#/components/schemas/SeriesEditor'
                required: true
            responses:
                "200":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Series'
                    description: Successful response
            summary: Update a series
            tags:
                - series
...
    /v1/seriesList:
        get:
            operationId: listSeries
            parameters:
                - in: query
                  name: Name
                  schema:
                    type: string
                - in: query
                  name: Slug
                  schema:
                    type: string
                - in: query
                  name: CreatedBefore
                  schema:
                    type: string
                - in: query
                  name: CreatedAfter
                  schema:
                    type: string
                - explode: true
                  in: query
                  name: SortBy
                  schema:
                    items:
                        type: string
                    type: array
                  style: form
                - description: Page number for pagination
                  in: query
                  name: page
                  schema:
                    default: 1
                    type: integer
            responses:
                "200":
                    content:
                        application/json:
                            schema:
                                items:
                                    $ref: '#/components/schemas/SeriesPartial'
                                type: array
                    description: Successful response
            summary: List series
            tags:
                - series
...
    /v1/user:
        get:
            operationId: getUser
            parameters:
                - in: query
                  name: id
                  required: true
                  schema:
                    type: integer
            responses:
                "200":
                    content:
                        application/json: {}
                    description: Successful response
            summary: Get a user account (admin)
            tags:
                - users
        post:
            operationId: updateUser
            requestBody:
                content:
                    application/json:
                        schema:
                            properties:
                                disabled:
                                    type: boolean
                                displayName:
                                    type: string
                                id:
                                    type: integer
                                password:
                                    type: string
                                role:
                                    type: string
                                scopeGroupId:
                                    nullable: true
                                    type: integer
                                username:
                                    type: string
                            type: object
                    application/x-www-form-urlencoded:
                        schema:
                            properties:
                                disabled:
                                    type: boolean
                                displayName:
                                    type: string
                                id:
                                    type: integer
                                password:
                                    type: string
                                role:
                                    type: string
                                scopeGroupId:
                                    nullable: true
                                    type: integer
                                username:
                                    type: string
                            type: object
                required: true
            responses:
                "200":
                    content:
                        application/json: {}
                    description: Successful response
            summary: Update a user account (admin)
            tags:
                - users
...
servers:
    - description: API Version 1
      url: /v1
...